refactor(api): consolidate GEO runtime bridge into runtime/geo
Move runGeoEffect and runRemoteGeoEffect from utils/geo-effect.ts to
runtime/geo.ts per Effect migration Phase 5. Update all GEO route imports
and remove the deprecated utils shim.
