+++ apps/api/src/routes/geo-settings.ts
@@ -13,8 +13,8 @@ import {
   GEO_COMMON_ERROR_RESPONSES,
   GEO_OPENAPI_TAG,
 } from "../constants/geo-openapi";
+import { runGeoEffect } from "../runtime/geo";
 import { findGeoSelectionError, geoErrorResponse } from "../utils/geo";
-import { runGeoEffect } from "../utils/geo-effect";
 import { createOpenApiApp } from "../utils/openapi-app";
 
 export const geoSettingsRoutes = createOpenApiApp();
