+++ apps/api/src/routes/geo-traffic.ts
@@ -34,8 +34,8 @@ import {
   GEO_COMMON_ERROR_RESPONSES,
   GEO_OPENAPI_TAG,
 } from "../constants/geo-openapi";
+import { runGeoEffect } from "../runtime/geo";
 import { geoErrorResponse } from "../utils/geo";
-import { runGeoEffect } from "../utils/geo-effect";
 import { logError } from "../utils/logging";
 import { createOpenApiApp } from "../utils/openapi-app";
