+++ apps/api/src/routes/posts.ts
@@ -32,6 +32,7 @@ import {
 import { and, count, eq, inArray, sql } from "drizzle-orm";
 import { nanoid } from "nanoid";
 
+import { runGeoEffect } from "../runtime/geo";
 import { addActiveGeneration } from "../utils/active-generations";
 import { getOrganizationId } from "../utils/auth";
 import {
@@ -42,7 +43,6 @@ import {
   resolveRequestedRepositoryIds,
   triggerContentGenerationWorkflow,
 } from "../utils/content-generation";
-import { runGeoEffect } from "../utils/geo-effect";
 import {
   extractTitleFromMarkdown,
   renderMarkdownToHtml,
